+++ packages/atproto/lib/src/labels/labels_service.dart
@@ -106,7 +106,7 @@ final class _LabelsService extends ATProtoBaseService implements LabelsService {
           'cursor': cursor,
         },
         to: LabelsByQuery.fromJson,
-        userContext: core.UserContext.anonymousOnly,
+        authType: core.AuthType.anonymous,
       );
 
   @override
@@ -119,7 +119,7 @@ final class _LabelsService extends ATProtoBaseService implements LabelsService {
             parameters: {
               'cursor': cursor,
             },
-            userContext: core.UserContext.anonymousOnly,
+            authType: core.AuthType.anonymous,
             to: SubscribedLabel.fromJson,
             adaptor: toSubscribedLabel,
           );