+++ packages/xrpc/lib/src/const.dart
@@ -2,7 +2,5 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
-enum UserContext {
-  anonymousOnly,
-  authRequired,
-}
+/// The default service to communicate.
+const defaultService = 'bsky.social';